Add camera vertical FOV parameter to renderer
Introduced a camera_angle_y field to the Camera struct to represent the vertical field of view in radians. Updated SampleRenderer to use this parameter instead of a hardcoded value, and ensured it is set appropriately in both depthMap and example04_firstTriangleMesh modules. Also updated the model file path in depthMap/main.cpp.